NONAME-133 added tooltip and modified account page
